+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 4.0.0 - 2020-01-08
+
+### Changed
+
+* isCamelCase is now replaced by caseStrategy=CamelCase
+
+### Added
+
+* Added caseStrategy that accept CamelCase | PascalCase | SnakeCase
+
 ## 3.5.0 - 2020-01-13
 
 ### Changed